// pipeline-cli/src/database/types/tables.ts
export enum SoftwareType {
  CAPTURE = 'capture',
  PROCESSING = 'processing',
  POST_PROCESSING = 'post_processing',
  ANALYSIS = 'analysis'
}
// ...
export interface Software {
  id: number;
  name: string;
  software_type: SoftwareType;
  version?: string;
  software_metadata?: SoftwareMetadata;
}

export interface MediaSoftware {
  id: number;
  file_id: number;
  software_id: number;
  usage_role: string;
  usage_metadata?: SoftwareUsageMetadata;
}